requirements.txt
django==4.2.3
pillow==10.0.0